+++ provdbconnector/__init__.py
@@ -1,3 +1,4 @@
 from provdbconnector.provapi import ProvApi
 from provdbconnector.db_adapters import BaseAdapter
 from provdbconnector.db_adapters.neo4j import Neo4jAdapter, NEO4J_USER, NEO4J_PASS, NEO4J_HOST, NEO4J_HTTP_PORT, NEO4J_BOLT_PORT
+from provdbconnector.tests import  *
