+++ provdbconnector/tests/__init__.py
@@ -1,0 +1,2 @@
+from provdbconnector.tests.db_adapters.test_baseadapter import AdapterTestTemplate
+from provdbconnector.tests.test_provapi import ProvApiTestTemplate
