+++ provdbconnector/tests/test_provapi.py
@@ -1,14 +1,14 @@
-import os
 import unittest
 from uuid import UUID
-import tests.examples as examples
-from io import StringIO
+
 import pkg_resources
 from prov.model import  ProvDocument
+
+from provdbconnector.tests import examples as examples
 from provdbconnector import ProvApi
-from provdbconnector.db_adapters.baseadapter import METADATA_KEY_TYPE_MAP,METADATA_KEY_PROV_TYPE,METADATA_KEY_IDENTIFIER,METADATA_KEY_NAMESPACES
 from provdbconnector.db_adapters import InvalidOptionsException
-from provdbconnector.db_adapters import Neo4jAdapter, NEO4J_USER, NEO4J_PASS, NEO4J_HOST, NEO4J_BOLT_PORT, NEO4J_HTTP_PORT
+from provdbconnector.db_adapters import Neo4jAdapter, NEO4J_USER, NEO4J_PASS, NEO4J_HOST, NEO4J_BOLT_PORT
+from provdbconnector.db_adapters.baseadapter import METADATA_KEY_TYPE_MAP,METADATA_KEY_PROV_TYPE,METADATA_KEY_IDENTIFIER,METADATA_KEY_NAMESPACES
 from provdbconnector.provapi import NoDataBaseAdapterException,InvalidArgumentTypeException
 
 
@@ -110,9 +110,9 @@ class ProvApiTests(unittest.TestCase):
     def setUp(self):
 
         self.test_resources = {
-            'xml': {'package': 'provdbconnector', 'file': '../tests/resources/primer.provx'},
-            'json': {'package': 'provdbconnector', 'file': '../tests/resources/primer.json'},
-            'provn': {'package': 'provdbconnector', 'file': '../tests/resources/primer.provn'}
+            'xml': {'package': 'provdbconnector', 'file': '/tests/resources/primer.provx'},
+            'json': {'package': 'provdbconnector', 'file': '/tests/resources/primer.json'},
+            'provn': {'package': 'provdbconnector', 'file': '/tests/resources/primer.provn'}
         }
         self.test_prov_files = dict((key, pkg_resources.resource_stream(val['package'], val['file'])) for key, val in self.test_resources.items())
         self.authInfo = {"user_name": NEO4J_USER,
